requirements.txt
dj-database-url==0.5.0
Django==2.1
django-ckeditor==5.6.1
django-cleanup==0.3.1
django-crispy-forms==1.7.2
django-js-asset==1.1.0
gunicorn==19.9.0
python-decouple==3.1
pytz==2018.5
whitenoise==4.0
